Use date taken instead of date created
The script was using the file system's date created instead of the photo's date taken. The script now looks at the exif data of the photo to determine the date taken. If no exif data exists or the file isn't TIFF, JPEG, PNG, Webp, or HEIC, the script uses the folder structure as the date with the day and time set to 0. For instance, 2004/11/photo.bmp will become 2004/11/20041100_1.bmp.